+++ cmd/relayproxy/proxynotifier/notifier_websocket.go
@@ -1,15 +1,15 @@
 package proxynotifier
 
 import (
-	"github.com/thomaspoignant/go-feature-flag/cmd/relayproxy/service"
+	"github.com/thomaspoignant/go-feature-flag/cmd/relayproxy/service/stream"
 	"github.com/thomaspoignant/go-feature-flag/notifier"
 )
 
 type notifierWebsocket struct {
-	websocketService service.WebsocketService
+	websocketService stream.WebsocketService
 }
 
-func NewNotifierWebsocket(websocketService service.WebsocketService) notifier.Notifier {
+func NewNotifierWebsocket(websocketService stream.WebsocketService) notifier.Notifier {
 	return &notifierWebsocket{
 		websocketService: websocketService,
 	}
