+++ cmd/relayproxy/service/stream/sse.go
@@ -1,4 +1,4 @@
-package service
+package stream
 
 import (
 	"encoding/json"
