+++ cmd/relayproxy/service/stream/websocket.go
@@ -1,4 +1,4 @@
-package service
+package stream
 
 import (
 	"context"
